Simplify get_page_title()
This is supercool!!  Congratulations on a great idea, well executed.

At the end, you cycle thru the pagenames 5 times in a row.  It would be better to cycle thru the list just once in a for loop doing the complete conversion on each element.  Or if you like list comprehensions (and who doesn't!) then you could create a `pagename_cleanup()` function to do it all in a single list comprehension.