HUGE improvement to speed in node highlighting
Now traceback highlighting is almost instant, even in huge networks with hundreds of nodes. Now, when a path is highlighted, the nodes and edges to be highlighted are stored in a global variable. When the time comes to reset node and edges color, only these nodes and edges are reset