Implement style and restore README
Now most of the style for the menu is in place. It is intentionally styled to look similar to the iOS UINavigationController, as is seen in the iOS "Settings" app. Also restore the README to the menu so that it is useful again. If you open the "help" menu and press "README", you will see the README along with a floating "back" button. The "About" and "Controls" pages have a heading that says "Not yet implemented" and a button that goes back to the main menu page.