Add deserialization methods
This does *not* currently reconstruct "floating edges," or edges that are not between a node and its parent. Floating edges only occur at the intersections between two separate graphs (only in large networks with multiple start nodes). This is pretty much an edge case (no pun intended 😛

This is also a step towards implementing #18