New “go” button behavior (Closes #34)
Now, instead of completely resetting the network, clicking “go” simply
makes the set of start nodes equal the set of pages entered in the
input field (it removes the items that were deleted from the input field
and adds the additional items that were entered). Since nodes are locked
when they are expanded, we never remove nodes that are integral to the
structure of the graph.